Remove test for a method from StatsBase (#233)
The code is testing that `counts` with two empty input arrays produces
an `ArgumentError`, which is failing because it actually produces a
`MethodError` from a Base function deeper in the call stack. The method
in question here is defined in StatsBase, not in this package, so it's
unclear why this package is testing it to begin with.